rename Map.js to Mapbox.js
